+++ karma.conf.js
@@ -1,5 +1,4 @@
 const webpackConfig = require('./webpack.dev.js');
-process.env.CHROME_BIN = require('puppeteer').executablePath();
 
 module.exports = (config) => {
   config.set({