+++ src/facades/events.js
@@ -1,0 +1,21 @@
+import helpers from '../helpers';
+
+const { error } = console;
+
+/**
+ * Sends a GET request to a given url.
+ *
+ * @param {string} url - The url to use
+ * in the request.
+ */
+export async function send(url) {
+  try {
+    if (url) {
+      helpers.ajax({
+        url,
+      });
+    }
+  } catch (err) {
+    error(err);
+  }
+}