+++ src/index.js
@@ -1,3 +1,2 @@
-export * from './pageRecommendations';
-export * from './pagesAndEventsRecommendations';
-export * from './productsRecommendations';
+export { default as pages } from './pages';
+export { default as events } from './events';